fix:added entitt list modal
